Fix exception from tell.js with empty db
If you start up nerdie and post a command for the tell.js
plugin when the tiny db for tell is empty, an exception is
thrown stating "Error: No records.". A string was hardcoded
to rethrow the error if it did not match exactly "No Records".
Since this did not match exactly, the exception was thrown anyway.
I changed this to a case-insensitive regex to match "no records"
to match even if the error message changes slightly.